refactor: move to internal package project layout
