+++ .goreleaser.yml
@@ -13,7 +13,7 @@ gomod:
   proxy: true
 
 builds:
-  - main: ./cmd/deepclean
+  - main: .
     env:
       - CGO_ENABLED=0
     goos:
