+++ internal/deepclean/results.go
