+++ internal/deepclean/scan.go
