+++ internal/deepclean/scan_test.go
